Implement file loading using Godot's File class
This allows us to load kernel, bios, and block device files in
exported projects.

In exported projects the files in 'res://' will typically be in a
compressed PCK file so loading the file doesn't work using the
C99/Posix FILE functions, but Godot's built-in File class is PCK
aware and can read files inside the PCK file.

However, the PCK file is read-only so give a warning when opening
'res://' files 'Read Write' when running from the editor and throw an
error when trying to open them 'Read Write' from an exported project.